requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-api-python-client==2.118.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
textblob==0.17.1
requests==2.31.0
